requirements.txt
aioredis==1.2.0
amqp==2.4.2
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
backcall==0.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
daphne==2.3.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2
django-cors-headers==2.5.2
django-filter==2.1.0
django-multiselectfield==0.1.8
django-rest-auth==0.9.5
djangorestframework==3.9.4
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
kombu==4.5.0
Markdown==3.1.1
msgpack==0.6.1
oauthlib==3.0.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.8.2
ptyprocess==0.6.0
Pygments==2.3.1
PyHamcrest==1.9.0
pyserial==3.4
python3-openid==3.1.0
pytz==2019.1
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
Twisted==19.2.0
txaio==18.8.1
urllib3==1.24.2
vine==1.3.0
wcwidth==0.1.7
zope.interface==4.6.0
